requirements.txt
flask==2.0.1
torch==1.12.1
torchvision==0.13.1
numpy==1.21.6
opencv-python==4.5.5.64
pillow==8.4.0
scikit-learn==1.0.2
matplotlib